Refactor token queue handling
