+++ pkg/service/service.go
@@ -25,8 +25,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
-	"time"
 
 	"github.com/ZaparooProject/zaparoo-core/pkg/assets"
 	"github.com/ZaparooProject/zaparoo-core/pkg/database/mediadb"
@@ -37,217 +35,13 @@ import (
 	"github.com/ZaparooProject/zaparoo-core/pkg/service/playlists"
 	"github.com/ZaparooProject/zaparoo-core/pkg/service/tokens"
 
-	"golang.org/x/exp/slices"
-
 	"github.com/ZaparooProject/zaparoo-core/pkg/config"
 	"github.com/ZaparooProject/zaparoo-core/pkg/database"
 	"github.com/ZaparooProject/zaparoo-core/pkg/groovyproxy"
 	"github.com/ZaparooProject/zaparoo-core/pkg/platforms"
 	"github.com/ZaparooProject/zaparoo-core/pkg/service/state"
-	"github.com/ZaparooProject/zaparoo-core/pkg/zapscript"
 	"github.com/rs/zerolog/log"
 )
-
-func inExitGameBlocklist(
-	cfg *config.Instance,
-	state *state.State,
-) bool {
-	var blocklist []string
-	for _, v := range cfg.ReadersScan().IgnoreSystem {
-		blocklist = append(blocklist, strings.ToLower(v))
-	}
-	return slices.Contains(blocklist, strings.ToLower(state.ActiveMedia().SystemID))
-}
-
-func launchToken(
-	platform platforms.Platform,
-	cfg *config.Instance,
-	token tokens.Token,
-	db *database.Database,
-	lsq chan<- *tokens.Token,
-	plsc playlists.PlaylistController,
-) error {
-	text := token.Text
-
-	mappingText, mapped := getMapping(cfg, db, platform, token)
-	if mapped {
-		log.Info().Msgf("found mapping: %s", mappingText)
-		text = mappingText
-	}
-
-	if text == "" {
-		return fmt.Errorf("no ZapScript in token")
-	}
-
-	log.Info().Msgf("launching ZapScript: %s", text)
-	cmds := strings.Split(text, "||")
-
-	pls := plsc.Active
-
-	for i, cmd := range cmds {
-		result, err := zapscript.LaunchToken(
-			platform,
-			cfg,
-			playlists.PlaylistController{
-				Active: pls,
-				Queue:  plsc.Queue,
-			},
-			token,
-			cmd,
-			len(cmds),
-			i,
-			db,
-		)
-		if err != nil {
-			return err
-		}
-
-		if result.MediaChanged && !token.FromAPI {
-			log.Debug().Any("token", token).Msg("media changed, updating token")
-			log.Info().Msgf("current media launched set to: %s", token.UID)
-			lsq <- &token
-		}
-
-		if result.PlaylistChanged {
-			pls = result.Playlist
-		}
-	}
-
-	return nil
-}
-
-func processTokenQueue(
-	platform platforms.Platform,
-	cfg *config.Instance,
-	st *state.State,
-	itq <-chan tokens.Token,
-	db *database.Database,
-	lsq chan<- *tokens.Token,
-	plq chan *playlists.Playlist,
-) {
-	for {
-		select {
-		case pls := <-plq:
-			activePlaylist := st.GetActivePlaylist()
-			launchPlaylistMedia := func() {
-				t := tokens.Token{
-					Text:     pls.Current().Path,
-					ScanTime: time.Now(),
-					Source:   tokens.SourcePlaylist,
-				}
-				plsc := playlists.PlaylistController{
-					Active: activePlaylist,
-					Queue:  plq,
-				}
-
-				err := launchToken(platform, cfg, t, db, lsq, plsc)
-				if err != nil {
-					log.Error().Err(err).Msgf("error launching token")
-				}
-
-				he := database.HistoryEntry{
-					Time:       t.ScanTime,
-					Type:       t.Type,
-					TokenID:    t.UID,
-					TokenValue: t.Text,
-					TokenData:  t.Data,
-				}
-				he.Success = err == nil
-				err = db.UserDB.AddHistory(he)
-				if err != nil {
-					log.Error().Err(err).Msgf("error adding history")
-				}
-			}
-
-			if pls == nil {
-				// playlist is cleared
-				if activePlaylist != nil {
-					log.Info().Msg("clearing playlist")
-				}
-				st.SetActivePlaylist(nil)
-				continue
-			} else if activePlaylist == nil {
-				// new playlist loaded
-				st.SetActivePlaylist(pls)
-				if pls.Playing {
-					log.Info().Any("pls", pls).Msg("setting new playlist, launching token")
-					go launchPlaylistMedia()
-				} else {
-					log.Info().Any("pls", pls).Msg("setting new playlist")
-				}
-				continue
-			} else {
-				// active playlist updated
-				if pls.Current() == activePlaylist.Current() &&
-					pls.Playing == activePlaylist.Playing {
-					log.Debug().Msg("playlist current token unchanged, skipping")
-					continue
-				}
-
-				st.SetActivePlaylist(pls)
-				if pls.Playing {
-					log.Info().Any("pls", pls).Msg("updating playlist, launching token")
-					go launchPlaylistMedia()
-				} else {
-					log.Info().Any("pls", pls).Msg("updating playlist")
-				}
-				continue
-			}
-		case t := <-itq:
-			// TODO: change this channel to send a token pointer or something
-			if t.ScanTime.IsZero() {
-				// ignore empty tokens
-				continue
-			}
-
-			log.Info().Msgf("processing token: %v", t)
-
-			err := platform.ScanHook(t)
-			if err != nil {
-				log.Error().Err(err).Msgf("error writing tmp scan result")
-			}
-
-			he := database.HistoryEntry{
-				Time:       t.ScanTime,
-				Type:       t.Type,
-				TokenID:    t.UID,
-				TokenValue: t.Text,
-				TokenData:  t.Data,
-			}
-
-			if !st.RunZapScriptEnabled() {
-				log.Debug().Msg("ZapScript disabled, skipping run")
-				err = db.UserDB.AddHistory(he)
-				if err != nil {
-					log.Error().Err(err).Msgf("error adding history")
-				}
-				continue
-			}
-
-			// launch tokens in separate thread
-			go func() {
-				plsc := playlists.PlaylistController{
-					Active: st.GetActivePlaylist(),
-					Queue:  plq,
-				}
-
-				err = launchToken(platform, cfg, t, db, lsq, plsc)
-				if err != nil {
-					log.Error().Err(err).Msgf("error launching token")
-				}
-
-				he.Success = err == nil
-				err = db.UserDB.AddHistory(he)
-				if err != nil {
-					log.Error().Err(err).Msgf("error adding history")
-				}
-			}()
-		case <-st.GetContext().Done():
-			log.Debug().Msg("Exiting Service worker via context cancellation")
-			break
-		}
-	}
-}
 
 func setupEnvironment(pl platforms.Platform) error {
 	if _, ok := utils.HasUserDir(); ok {
